requirements.txt
# Base
pytz==2019.3
python-slugify==4.0.0
Pillow==7.0.0
django-cleanup==4.0.0
psycopg2-binary==2.8.5

# Django
django==3.0.3

# Environment
django-environ==0.4.5

# Django REST Framework
djangorestframework==3.11.1
django-filter==2.2.0

# JWT
pyjwt>=1.7.1

# Passwords security
argon2-cffi==19.2.0

# Static files
whitenoise==5.0.1

# Celery
vine==1.3.0
redis>=3.2.0
django-redis==4.10.0
celery==4.4.7
flower==0.9.7

# Tools
django-extensions==2.2.9
django-utils-six==2.0
launchpadlib==1.10.13

# Debugging
ipdb==0.11


# Testing
# mypy==0.650
pytest==6.0.0
pytest-sugar==0.9.4
pytest-django==4.1.0
factory-boy==3.2.0
mypy==0.790
attrs==20.3.0

# Code quality
flake8==3.8.4

#Stripe
stripe

Faker==4.1.1
django-webpack-loader==0.6.0

gunicorn==20.0.4
daphne==3.0.1

# Static files
django-storages[boto3]==1.9.1

# Email
django-anymail[sendgrid]==7.2.1

# Video
moviepy==1.0.3

# Channels
channels==3.0.2

django-cors-headers==3.6.0

requests==2.25.1

channels-redis==3.2.0

autopep8==1.5.4

pycodestyle==2.6.0

# GeoIp
geoip2==4.1.0

# Geopy
geopy==2.1.0

# Currency by country
ccy==1.2.2

django-money[exchange]==1.3

forex-python==1.5

# Paypal
paypal-payouts-sdk==1.0.0

# Coinbase commerce
coinbase-commerce==1.0.1

